--- packages/backend-next/package.json
@@ -1,70 +1,0 @@
-{
-  "name": "example-backend-next",
-  "version": "0.0.17",
-  "main": "dist/index.cjs.js",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "private": true,
-  "backstage": {
-    "role": "backend"
-  },
-  "homepage": "https://backstage.io",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/backstage",
-    "directory": "packages/backend-next"
-  },
-  "keywords": [
-    "backstage"
-  ],
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test",
-    "clean": "backstage-cli package clean"
-  },
-  "dependencies": {
-    "@backstage/backend-defaults": "workspace:^",
-    "@backstage/backend-plugin-api": "workspace:^",
-    "@backstage/backend-tasks": "workspace:^",
-    "@backstage/plugin-adr-backend": "workspace:^",
-    "@backstage/plugin-app-backend": "workspace:^",
-    "@backstage/plugin-auth-node": "workspace:^",
-    "@backstage/plugin-azure-devops-backend": "workspace:^",
-    "@backstage/plugin-badges-backend": "workspace:^",
-    "@backstage/plugin-catalog-backend": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-backstage-openapi": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-openapi": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "workspace:^",
-    "@backstage/plugin-catalog-backend-module-unprocessed": "workspace:^",
-    "@backstage/plugin-devtools-backend": "workspace:^",
-    "@backstage/plugin-entity-feedback-backend": "workspace:^",
-    "@backstage/plugin-jenkins-backend": "workspace:^",
-    "@backstage/plugin-kubernetes-backend": "workspace:^",
-    "@backstage/plugin-lighthouse-backend": "workspace:^",
-    "@backstage/plugin-linguist-backend": "workspace:^",
-    "@backstage/plugin-nomad-backend": "workspace:^",
-    "@backstage/plugin-permission-backend": "workspace:^",
-    "@backstage/plugin-permission-backend-module-allow-all-policy": "workspace:^",
-    "@backstage/plugin-permission-common": "workspace:^",
-    "@backstage/plugin-permission-node": "workspace:^",
-    "@backstage/plugin-playlist-backend": "workspace:^",
-    "@backstage/plugin-proxy-backend": "workspace:^",
-    "@backstage/plugin-scaffolder-backend": "workspace:^",
-    "@backstage/plugin-search-backend": "workspace:^",
-    "@backstage/plugin-search-backend-module-catalog": "workspace:^",
-    "@backstage/plugin-search-backend-module-explore": "workspace:^",
-    "@backstage/plugin-search-backend-module-techdocs": "workspace:^",
-    "@backstage/plugin-search-backend-node": "workspace:^",
-    "@backstage/plugin-sonarqube-backend": "workspace:^",
-    "@backstage/plugin-techdocs-backend": "workspace:^",
-    "@backstage/plugin-todo-backend": "workspace:^"
-  },
-  "devDependencies": {
-    "@backstage/cli": "workspace:^"
-  },
-  "files": [
-    "dist"
-  ]
-}
