+++ packages/backend-old/README.md
@@ -1,0 +1,5 @@
+# example-backend-next
+
+This is an example backend for the old Backstage backend system.
+
+Do not use this in your own projects.
