+++ packages/backend-old/catalog-info.yaml
@@ -1,8 +1,8 @@
 apiVersion: backstage.io/v1alpha1
 kind: Component
 metadata:
-  name: example-backend-next
-  title: example-backend-next
+  name: example-backend-old
+  title: example-backend-old
 spec:
   lifecycle: experimental
   type: backstage-backend
