+++ packages/backend-old/src/plugins/lighthouse.ts
