+++ packages/backend/package.json
@@ -22,87 +22,47 @@
     "build": "backstage-cli package build",
     "lint": "backstage-cli package lint",
     "test": "backstage-cli package test",
-    "clean": "backstage-cli package clean",
-    "build-image": "docker build ../.. -f Dockerfile --tag example-backend"
+    "clean": "backstage-cli package clean"
   },
   "dependencies": {
-    "@backstage/backend-common": "workspace:^",
+    "@backstage/backend-defaults": "workspace:^",
+    "@backstage/backend-plugin-api": "workspace:^",
     "@backstage/backend-tasks": "workspace:^",
-    "@backstage/catalog-client": "workspace:^",
-    "@backstage/catalog-model": "workspace:^",
-    "@backstage/config": "workspace:^",
-    "@backstage/integration": "workspace:^",
     "@backstage/plugin-adr-backend": "workspace:^",
     "@backstage/plugin-app-backend": "workspace:^",
-    "@backstage/plugin-auth-backend": "workspace:^",
     "@backstage/plugin-auth-node": "workspace:^",
     "@backstage/plugin-azure-devops-backend": "workspace:^",
-    "@backstage/plugin-azure-sites-backend": "workspace:^",
     "@backstage/plugin-badges-backend": "workspace:^",
     "@backstage/plugin-catalog-backend": "workspace:^",
+    "@backstage/plugin-catalog-backend-module-backstage-openapi": "workspace:^",
+    "@backstage/plugin-catalog-backend-module-openapi": "workspace:^",
     "@backstage/plugin-catalog-backend-module-scaffolder-entity-model": "workspace:^",
     "@backstage/plugin-catalog-backend-module-unprocessed": "workspace:^",
-    "@backstage/plugin-catalog-node": "workspace:^",
-    "@backstage/plugin-code-coverage-backend": "workspace:^",
     "@backstage/plugin-devtools-backend": "workspace:^",
     "@backstage/plugin-entity-feedback-backend": "workspace:^",
-    "@backstage/plugin-events-backend": "workspace:^",
-    "@backstage/plugin-events-node": "workspace:^",
-    "@backstage/plugin-explore-backend": "workspace:^",
-    "@backstage/plugin-graphql-backend": "workspace:^",
     "@backstage/plugin-jenkins-backend": "workspace:^",
-    "@backstage/plugin-kafka-backend": "workspace:^",
     "@backstage/plugin-kubernetes-backend": "workspace:^",
     "@backstage/plugin-lighthouse-backend": "workspace:^",
     "@backstage/plugin-linguist-backend": "workspace:^",
     "@backstage/plugin-nomad-backend": "workspace:^",
     "@backstage/plugin-permission-backend": "workspace:^",
+    "@backstage/plugin-permission-backend-module-allow-all-policy": "workspace:^",
     "@backstage/plugin-permission-common": "workspace:^",
     "@backstage/plugin-permission-node": "workspace:^",
     "@backstage/plugin-playlist-backend": "workspace:^",
     "@backstage/plugin-proxy-backend": "workspace:^",
-    "@backstage/plugin-rollbar-backend": "workspace:^",
     "@backstage/plugin-scaffolder-backend": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-confluence-to-markdown": "workspace:^",
-    "@backstage/plugin-scaffolder-backend-module-rails": "workspace:^",
     "@backstage/plugin-search-backend": "workspace:^",
     "@backstage/plugin-search-backend-module-catalog": "workspace:^",
-    "@backstage/plugin-search-backend-module-elasticsearch": "workspace:^",
     "@backstage/plugin-search-backend-module-explore": "workspace:^",
-    "@backstage/plugin-search-backend-module-pg": "workspace:^",
     "@backstage/plugin-search-backend-module-techdocs": "workspace:^",
     "@backstage/plugin-search-backend-node": "workspace:^",
-    "@backstage/plugin-search-common": "workspace:^",
-    "@backstage/plugin-tech-insights-backend": "workspace:^",
-    "@backstage/plugin-tech-insights-backend-module-jsonfc": "workspace:^",
-    "@backstage/plugin-tech-insights-node": "workspace:^",
+    "@backstage/plugin-sonarqube-backend": "workspace:^",
     "@backstage/plugin-techdocs-backend": "workspace:^",
-    "@backstage/plugin-todo-backend": "workspace:^",
-    "@gitbeaker/node": "^35.1.0",
-    "@octokit/rest": "^19.0.3",
-    "@opentelemetry/api": "^1.4.1",
-    "@opentelemetry/exporter-prometheus": "^0.45.0",
-    "@opentelemetry/sdk-metrics": "^1.13.0",
-    "azure-devops-node-api": "^11.0.1",
-    "better-sqlite3": "^9.0.0",
-    "dockerode": "^3.3.1",
-    "example-app": "link:../app",
-    "express": "^4.17.1",
-    "express-prom-bundle": "^6.3.6",
-    "express-promise-router": "^4.1.0",
-    "luxon": "^3.0.0",
-    "mysql2": "^2.2.5",
-    "pg": "^8.11.3",
-    "pg-connection-string": "^2.3.0",
-    "prom-client": "^14.0.1",
-    "winston": "^3.2.1"
+    "@backstage/plugin-todo-backend": "workspace:^"
   },
   "devDependencies": {
-    "@backstage/cli": "workspace:^",
-    "@types/dockerode": "^3.3.0",
-    "@types/express": "^4.17.6",
-    "@types/express-serve-static-core": "^4.17.5",
-    "@types/luxon": "^3.0.0"
+    "@backstage/cli": "workspace:^"
   },
   "files": [
     "dist"
